backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
Pillow>=10.0.0
opencv-python-headless>=4.9.0
numpy>=1.24.0
scikit-image>=0.22.0
transformers>=4.36.0
torch>=2.0.0
